+++ demo/button.gd
@@ -3,19 +3,32 @@ extends Button
 export(String) var scene
 export(float) var fade_out_speed = 1
 export(float) var fade_in_speed = 1
+export(String) var fade_out_pattern = "fade"
+export(String) var fade_in_pattern = "fade"
+export(float, 0, 1) var fade_out_smoothness = 0.1
+export(float, 0, 1) var fade_in_smoothness = 0.1
+export(bool) var fade_out_inverted = false
+export(bool) var fade_in_inverted = false
 export(Color) var color = Color(0, 0, 0)
 export(float) var timeout = 0
 export(bool) var clickable = false
-onready var scene_options = SceneManager.create_options(fade_out_speed, fade_in_speed, color, timeout, clickable)
+
+onready var fade_out_options = SceneManager.create_options(fade_out_speed, fade_out_pattern, fade_out_smoothness, fade_out_inverted)
+onready var fade_in_options = SceneManager.create_options(fade_in_speed, fade_in_pattern, fade_in_smoothness, fade_in_inverted)
+onready var general_options = SceneManager.create_general_options(color, timeout, clickable)
 
 func _ready() -> void:
-	var first_scene_options = SceneManager.create_options(0, 1, Color(1, 1, 1), 1, false)
-	SceneManager.show_first_scene(first_scene_options)
-	# code break happens if scene is not recognizable
-	SceneManager.validate_key(scene)
+	var fade_in_first_scene_options = SceneManager.create_options(1, "fade")
+	var first_scene_general_options = SceneManager.create_general_options(Color(0, 0, 0), 1, false)
+	SceneManager.show_first_scene(fade_in_first_scene_options, first_scene_general_options)
+	# code breaks if scene is not recognizable
+	SceneManager.validate_scene(scene)
+	# code breaks if pattern is not recognizable
+	SceneManager.validate_pattern(fade_out_pattern)
+	SceneManager.validate_pattern(fade_in_pattern)
 
 func _on_button_button_up():
-	SceneManager.change_scene(scene, scene_options)
+	SceneManager.change_scene(scene, fade_out_options, fade_in_options, general_options)
 
 func _on_reset_button_up():
 	SceneManager.reset_scene_manager()