+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ CHANGELOG
 UNRELEASED
 ----------
 
+* 🎉 feat: shader transition feature added, create_scene_instance function added, transition is so much more customizable
 * 🎉 feat: shader scene transition added
 
 1.3.0 (2022-03-18)